Fixed compile warnings about name shadowing

+++ src/common/base/NebulaKeyUtils.cpp
@@ -13,7 +13,7 @@ std::string NebulaKeyUtils::vertexKey(PartitionID partId, VertexID vId,
                                       TagID tagId, TagVersion tv) {
     constexpr uint32_t tagMask = 0xBFFFFFFF;
     tagId &= tagMask;
-    int32_t item = (partId << 8) | (NebulaKeyType::kData);
+    int32_t item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
 
     std::string key;
     key.reserve(kVertexLen);
@@ -33,7 +33,7 @@ std::string NebulaKeyUtils::edgeKey(PartitionID partId,
                                     EdgeVersion ev) {
     constexpr uint32_t edgeMask = 0x40000000;
     type |= edgeMask;
-    int32_t item = (partId << 8) | (NebulaKeyType::kData);
+    int32_t item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
 
     std::string key;
     key.reserve(kEdgeLen);
@@ -48,8 +48,8 @@ std::string NebulaKeyUtils::edgeKey(PartitionID partId,
 
 // static
 std::string NebulaKeyUtils::systemCommitKey(PartitionID partId) {
-    int32_t item = (partId << 8) | (NebulaKeyType::kSystem);
-    uint32_t type = NebulaSystemKeyType::kSystemCommit;
+    int32_t item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kSystem);
+    uint32_t type = static_cast<uint32_t>(NebulaSystemKeyType::kSystemCommit);
     std::string key;
     key.reserve(kSystemLen);
     key.append(reinterpret_cast<const char*>(&item), sizeof(PartitionID))
@@ -59,8 +59,8 @@ std::string NebulaKeyUtils::systemCommitKey(PartitionID partId) {
 
 // static
 std::string NebulaKeyUtils::systemPartKey(PartitionID partId) {
-    uint32_t item = (partId << 8) | (NebulaKeyType::kSystem);
-    uint32_t type = NebulaSystemKeyType::kSystemPart;
+    uint32_t item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kSystem);
+    uint32_t type = static_cast<uint32_t>(NebulaSystemKeyType::kSystemPart);
     std::string key;
     key.reserve(kSystemLen);
     key.append(reinterpret_cast<const char*>(&item), sizeof(PartitionID))
@@ -72,7 +72,7 @@ std::string NebulaKeyUtils::systemPartKey(PartitionID partId) {
 std::string NebulaKeyUtils::uuidKey(PartitionID partId, const folly::StringPiece& name) {
     std::string key;
     key.reserve(sizeof(PartitionID) + name.size());
-    int32_t item = (partId << 8) | (NebulaKeyType::kUUID);
+    int32_t item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kUUID);
     key.append(reinterpret_cast<const char*>(&item), sizeof(int32_t))
        .append(name.data(), name.size());
     return key;
@@ -82,7 +82,7 @@ std::string NebulaKeyUtils::uuidKey(PartitionID partId, const folly::StringPiece
 std::string NebulaKeyUtils::vertexPrefix(PartitionID partId, VertexID vId, TagID tagId) {
     constexpr uint32_t tagMask = 0xBFFFFFFF;
     tagId &= tagMask;
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
 
     std::string key;
     key.reserve(kVertexLen);
@@ -96,7 +96,7 @@ std::string NebulaKeyUtils::vertexPrefix(PartitionID partId, VertexID vId, TagID
 std::string NebulaKeyUtils::edgePrefix(PartitionID partId, VertexID srcId, EdgeType type) {
     constexpr uint32_t edgeMask = 0x40000000;
     type |= edgeMask;
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
 
     std::string key;
     key.reserve(sizeof(PartitionID) + sizeof(VertexID) + sizeof(EdgeType));
@@ -108,7 +108,7 @@ std::string NebulaKeyUtils::edgePrefix(PartitionID partId, VertexID srcId, EdgeT
 
 // static
 std::string NebulaKeyUtils::prefix(PartitionID partId) {
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
     std::string key;
     key.reserve(sizeof(PartitionID));
     key.append(reinterpret_cast<const char*>(&item), sizeof(PartitionID));
@@ -117,7 +117,7 @@ std::string NebulaKeyUtils::prefix(PartitionID partId) {
 
 // static
 std::string NebulaKeyUtils::vertexPrefix(PartitionID partId, VertexID vId) {
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
     std::string key;
     key.reserve(sizeof(PartitionID) + sizeof(VertexID));
     key.append(reinterpret_cast<const char*>(&item), sizeof(PartitionID))
@@ -127,7 +127,7 @@ std::string NebulaKeyUtils::vertexPrefix(PartitionID partId, VertexID vId) {
 
 // static
 std::string NebulaKeyUtils::edgePrefix(PartitionID partId, VertexID vId) {
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
     std::string key;
     key.reserve(sizeof(PartitionID) + sizeof(VertexID));
     key.append(reinterpret_cast<const char*>(&item), sizeof(PartitionID))
@@ -137,7 +137,7 @@ std::string NebulaKeyUtils::edgePrefix(PartitionID partId, VertexID vId) {
 
 // static
 std::string NebulaKeyUtils::systemPrefix() {
-    int8_t type = NebulaKeyType::kSystem;
+    int8_t type = static_cast<uint32_t>(NebulaKeyType::kSystem);
     std::string key;
     key.reserve(sizeof(int8_t));
     key.append(reinterpret_cast<const char*>(&type), sizeof(int8_t));
@@ -149,7 +149,7 @@ std::string NebulaKeyUtils::prefix(PartitionID partId, VertexID src, EdgeType ty
                                    EdgeRanking ranking, VertexID dst) {
     constexpr uint32_t edgeMask = 0x40000000;
     type |= edgeMask;
-    PartitionID item = (partId << 8) | (NebulaKeyType::kData);
+    PartitionID item = (partId << 8) | static_cast<uint32_t>(NebulaKeyType::kData);
 
     std::string key;
     key.reserve(sizeof(PartitionID) + sizeof(VertexID) + sizeof(EdgeType)
